+++ Servers/Common/vtkPVClientServerIdCollectionInformation.cxx
@@ -152,7 +152,7 @@ int vtkPVClientServerIdCollectionInformation
 //----------------------------------------------------------------------------
 int vtkPVClientServerIdCollectionInformation::GetLength()
 {
-  return this->ClientServerIds->size();
+  return static_cast<int>(this->ClientServerIds->size());
 }
 
 //----------------------------------------------------------------------------
