+++ Servers/Common/vtkPVFileInformation.cxx
@@ -436,7 +436,7 @@ void vtkPVFileInformation::CopyFromObject(vtkObject* object)
     path.replace(idx, 1, 1, '\\');
     }
   
-  int len = path.size();
+  int len = static_cast<int>(path.size());
   if(len > 4 && path.compare(len-4, 4, ".lnk") == 0)
     {
     WIN32_FIND_DATA data;
